Aligns packages with namespaces for android-platform
