+++ features/facts/src/main/java/io/dotanuki/norris/features/facts/ui/FactsActivity.kt
@@ -5,9 +5,9 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.lifecycleScope
 import androidx.lifecycle.repeatOnLifecycle
-import io.dotanuki.norris.common.android.selfBind
 import io.dotanuki.norris.features.facts.presentation.FactsUserInteraction
 import io.dotanuki.norris.features.facts.presentation.FactsViewModel
+import io.dotanuki.platform.android.core.helpers.selfBind
 import kotlinx.coroutines.launch
 import org.kodein.di.DIAware
 import org.kodein.di.instance
