--- platform/android/core-helpers/src/main/java/io/dotanuki/norris/common/android/ActivityExtensions.kt
@@ -1,30 +1,0 @@
-package io.dotanuki.norris.common.android
-
-import android.view.LayoutInflater
-import androidx.appcompat.app.AppCompatActivity
-import androidx.fragment.app.FragmentActivity
-import androidx.viewbinding.ViewBinding
-import io.dotanuki.platform.jvm.core.kodein.KodeinTags
-import org.kodein.di.DI
-import org.kodein.di.DIAware
-import org.kodein.di.bind
-import org.kodein.di.provider
-
-fun AppCompatActivity.selfBind(bindings: DI.MainBuilder.() -> Unit = {}) = DI.lazy {
-
-    val parentContainer = (applicationContext as DIAware).di
-    extend(parentContainer)
-
-    bind<FragmentActivity>(tag = KodeinTags.hostActivity) with provider {
-        this@selfBind
-    }
-
-    bindings.invoke(this)
-}
-
-inline fun <T : ViewBinding> AppCompatActivity.viewBinding(
-    crossinline binder: (LayoutInflater) -> T
-) =
-    lazy(LazyThreadSafetyMode.NONE) {
-        binder.invoke(layoutInflater)
-    }
