--- platform/android/core-navigator/src/main/java/io/dotanuki/norris/navigator/Screen.kt
@@ -1,11 +1,0 @@
-package io.dotanuki.norris.navigator
-
-sealed class Screen {
-    object FactsList : Screen()
-    object SearchQuery : Screen()
-
-    override fun toString() = when (this) {
-        FactsList -> "Facts Screen"
-        SearchQuery -> "SearchQuery Query Screen"
-    }
-}
