--- platform/android/core-navigator/src/main/java/io/dotanuki/norris/navigator/UnsupportedNavigation.kt
@@ -1,5 +1,0 @@
-package io.dotanuki.norris.navigator
-
-data class UnsupportedNavigation(val destination: Screen) : RuntimeException(
-    "Cannot navigate to this destination -> $destination"
-)
