+++ platform/android/core-navigator/src/main/java/io/dotanuki/platform/android/core/navigator/UnsupportedNavigation.kt
@@ -1,0 +1,5 @@
+package io.dotanuki.platform.android.core.navigator
+
+data class UnsupportedNavigation(val destination: Screen) : RuntimeException(
+    "Cannot navigate to this destination -> $destination"
+)
