--- platform/android/core-persistance/src/main/java/io/dotanuki/norris/persistance/AppPreferencesWrapper.kt
@@ -1,16 +1,0 @@
-package io.dotanuki.norris.persistance
-
-import android.app.Application
-import android.content.Context.MODE_PRIVATE
-import android.content.SharedPreferences
-
-class AppPreferencesWrapper(private val app: Application) {
-
-    val preferences: SharedPreferences by lazy {
-        app.getSharedPreferences(PREFS_FILE, MODE_PRIVATE)
-    }
-
-    private companion object {
-        const val PREFS_FILE = "last-searches"
-    }
-}
