--- platform/android/core-persistance/src/main/java/io/dotanuki/norris/persistance/LocalStorage.kt
@@ -1,10 +1,0 @@
-package io.dotanuki.norris.persistance
-
-interface LocalStorage {
-
-    suspend fun lastSearches(): List<String>
-
-    fun registerNewSearch(term: String)
-
-    fun destroy()
-}
