+++ platform/android/core-persistance/src/main/java/io/dotanuki/platform/android/core/persistance/LocalStorage.kt
@@ -1,0 +1,10 @@
+package io.dotanuki.platform.android.core.persistance
+
+interface LocalStorage {
+
+    suspend fun lastSearches(): List<String>
+
+    fun registerNewSearch(term: String)
+
+    fun destroy()
+}
