--- platform/android/testing-screenshots/src/main/java/io/dotanuki/testing/screenshots/ScreenshotDriver.kt
@@ -1,10 +1,0 @@
-package io.dotanuki.testing.screenshots
-
-import android.app.Activity
-
-interface ScreenshotDriver<A : Activity, State> {
-
-    fun beforeCapturing(target: A, state: State)
-
-    fun imageName(state: State): String
-}
